remove resource base class
